feat: add content parsing and handling components
