+++ src/components/Page/Post/Content/handlers/CodePenHandler.tsx
@@ -1,14 +1,10 @@
 import { Text } from 'html-react-parser';
-import type { HandlerFunction } from '../types';
+import type { HandlerFunction } from './types';
 
 /**
- * CodePen埋め込みのハンドラー
  * CodePenのiframeを処理し、ハイドレーション時のエスケープ文字不一致を防止する
- * @param domNode - 処理対象のDOMノード
- * @returns 処理済みのDOMノードまたはundefined
  */
 export const handleCodePen: HandlerFunction = (domNode) => {
-  // CodePenのiframeを処理する
   if (
     domNode.tagName === 'iframe' &&
     typeof domNode.attribs?.src === 'string' &&
