+++ src/components/Page/Post/Content/handlers/index.ts
@@ -2,3 +2,5 @@ export { handleAlert } from './AlertHandler';
 export { handleAnchor } from './AnchorHandler';
 export { handleCodePen } from './CodePenHandler';
 export { handleLinkPreview } from './LinkPreviewHandler';
+export type { HandlerFunction } from './types';
+export { handleZoomImage } from './zoom-image/ZoomImageHandler';
