+++ src/components/Page/Post/Content/handlers/types.ts
@@ -1,0 +1,3 @@
+import type { Element } from 'html-react-parser';
+
+export type HandlerFunction = (domNode: Element) => React.ReactElement | null | undefined;
