--- src/components/Page/Post/Content/inject/index.ts
@@ -1,1 +1,0 @@
-export { parser } from './HtmlParser';
