+++ src/components/Page/Post/Content/parser/HTMLParser.tsx
@@ -1,0 +1,27 @@
+import reactHtmlParser, { Element, type HTMLReactParserOptions } from 'html-react-parser';
+import { handleAlert, handleAnchor, handleCodePen, handleLinkPreview, handleZoomImage } from '../handlers';
+
+const applyHandlers = (domNode: Element) => {
+  return (
+    handleAnchor(domNode) ||
+    handleAlert(domNode) ||
+    handleLinkPreview(domNode) ||
+    handleCodePen(domNode) ||
+    handleZoomImage(domNode) ||
+    domNode
+  );
+};
+
+const reactHtmlParserOptions: HTMLReactParserOptions = {
+  replace: (domNode) => {
+    if (!(domNode instanceof Element && domNode.attribs)) {
+      return domNode;
+    }
+
+    return applyHandlers(domNode);
+  },
+};
+
+export const parser = (content: string) => {
+  return reactHtmlParser(content, reactHtmlParserOptions);
+};
