+++ src/components/Page/Post/index.ts
@@ -1,4 +1,4 @@
-import Content from './Content';
+import { PostContent as Content } from './Content';
 import Edit from './Edit';
 import Header from './Header';
 import Note from './Note';
